worker/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements.txt requirements.in
#
agentarchives==0.9.0
    # via -r requirements.in
ammcpc==0.2.0
    # via -r requirements.in
asgiref==3.8.1
    # via django
bagit @ git+https://github.com/artefactual-labs/bagit-python.git@4b8fde73b4e631461bfd7add87e200500d40ca21
    # via -r requirements.in
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
clamd==1.0.2
    # via -r requirements.in
django==4.2.15
    # via -r requirements.in
django-autoslug==1.9.9
    # via -r requirements.in
django-tastypie==0.14.7
    # via -r requirements.in
gearman3 @ git+https://github.com/artefactual-labs/python-gearman.git@b68efc868c7a494dd6a2d2e820fb098a6da9f797
    # via -r requirements.in
idna==3.7
    # via requests
importlib-resources==6.4.0
    # via opf-fido
lxml==5.2.2
    # via
    #   -r requirements.in
    #   ammcpc
    #   metsrw
metsrw==0.5.1
    # via -r requirements.in
mysqlclient==2.2.4
    # via agentarchives
olefile==0.47
    # via opf-fido
opf-fido @ git+https://github.com/artefactual-labs/fido.git@564ceb8018a8650fe931cf20e6780ee008e60fca
    # via -r requirements.in
orjson==3.10.6
    # via -r requirements.in
prometheus-client==0.20.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via django-tastypie
python-mimeparse==1.6.0
    # via django-tastypie
requests==2.32.3
    # via
    #   -r requirements.in
    #   agentarchives
    #   opf-fido
six==1.16.0
    # via
    #   opf-fido
    #   python-dateutil
sqlparse==0.5.0
    # via django
unidecode==1.3.8
    # via -r requirements.in
urllib3==2.2.2
    # via requests
